refactor(api-bundle)!: drop deprecated ManageApiToken names
Remove the deprecated ManageApiTokenAuth attribute and ManageApiToken token
class (and the inverted-inheritance/class_alias shims) in favour of the
KubernetesServiceAccount* names only. The X-KBC-ManageApiToken header remains
fully supported inside KubernetesServiceAccountAuthenticator alongside the
X-Kubernetes-Authorization JWT header.

BREAKING CHANGE: consumers must use KubernetesServiceAccountAuth /
KubernetesServiceAccountToken and the new Security\KubernetesServiceAccount
namespace; release as a major version.
