+++ libs/api-bundle/README.md
@@ -103,13 +103,12 @@ class Controller {
 
 `KubernetesServiceAccountAuth` accepts both the Manage API token header (`X-KBC-ManageApiToken`)
 and the Kubernetes ServiceAccount JWT header (`X-Kubernetes-Authorization`); Connection resolves
-both to a Manage token, so `scopes`/`isSuperAdmin` checks are identical. The `ManageApiTokenAuth`
-attribute is a deprecated 1:1 alias kept for backwards compatibility — use
-`KubernetesServiceAccountAuth` instead.
+both to a Manage token, so `scopes`/`isSuperAdmin` checks are identical regardless of which header
+the request carries.
 
 To use individual authentication attributes, you need to install appropriate client package:
 * to use `StorageApiTokenAuth`, install `keboola/storage-api-client`
-* to use `KubernetesServiceAccountAuth` (or the deprecated `ManageApiTokenAuth`), install `keboola/kbc-manage-api-php-client`
+* to use `KubernetesServiceAccountAuth`, install `keboola/kbc-manage-api-php-client`
 
 > [!NOTE]
 > If you forget to install appropriate client, you will get exception like
